Move PatchOp class to separate source file.
- move declarations and implementations of PatchOp descendants to one file
- fix includes
